move type checks to be as part of spark parity tests
