+++ mleap-runtime/src/main/scala/ml/combust/mleap/runtime/transformer/feature/BucketedRandomProjectionLSH.scala
@@ -6,7 +6,6 @@ import ml.combust.mleap.runtime.function.UserDefinedFunction
 import ml.combust.mleap.runtime.transformer.{SimpleTransformer, Transformer}
 import ml.combust.mleap.tensor.Tensor
 import ml.combust.mleap.core.util.VectorConverters._
-import ml.combust.mleap.runtime.SchemaUtil
 
 /**
   * Created by hollinwilkins on 12/28/16.
@@ -15,6 +14,4 @@ case class BucketedRandomProjectionLSH(override val uid: String = Transformer.un
                                        override val shape: NodeShape,
                                        override val model: BucketedRandomProjectionLSHModel) extends SimpleTransformer {
   override val exec: UserDefinedFunction = (features: Tensor[Double]) => model(features)
-
-  SchemaUtil.asssertModelTypesMatchTransformerTypes(model, exec)
 }
